Clean up search land tickets flow
