+++ packages/manager/src/features/Support/SupportTickets/TicketList.tsx
@@ -1,71 +1,30 @@
 import { SupportTicket } from '@linode/api-v4/lib/support';
-import { compose } from 'ramda';
 import * as React from 'react';
-import Paper from 'src/components/core/Paper';
-import {
-  createStyles,
-  withStyles,
-  WithStyles,
-} from 'src/components/core/styles';
+import { compose } from 'recompose';
+import Hidden from 'src/components/core/Hidden';
 import TableBody from 'src/components/core/TableBody';
 import TableHead from 'src/components/core/TableHead';
 import { OrderByProps } from 'src/components/OrderBy';
 import Pagey, { PaginationProps } from 'src/components/Pagey';
 import PaginationFooter from 'src/components/PaginationFooter';
-import Table from 'src/components/Table';
-import TableCell from 'src/components/TableCell';
-import TableRow from 'src/components/TableRow';
+import Table from 'src/components/Table/Table_CMR';
+import TableCell from 'src/components/TableCell/TableCell_CMR';
+import TableRow from 'src/components/TableRow/TableRow_CMR';
 import TableRowEmptyState from 'src/components/TableRowEmptyState';
 import TableRowError from 'src/components/TableRowError';
 import TableRowLoading from 'src/components/TableRowLoading';
-import TableSortCell from 'src/components/TableSortCell';
+import TableSortCell from 'src/components/TableSortCell/TableSortCell_CMR';
 import TicketRow from './TicketRow';
 import { getTicketsPage } from './ticketUtils';
 
-interface Props extends PaginationProps<SupportTicket> {
+interface Props {
   filterStatus: 'open' | 'closed';
   newTicket?: SupportTicket;
 }
 
-type ClassNames =
-  | 'root'
-  | 'cellSubject'
-  | 'cellId'
-  | 'cellRegarding'
-  | 'cellCreated'
-  | 'cellUpdated'
-  | 'cellUpdatedBy';
-
-const styles = () =>
-  createStyles({
-    root: {},
-    cellSubject: {
-      width: '35%',
-      minWidth: 175,
-    },
-    cellId: {
-      width: '10%',
-    },
-    cellRegarding: {
-      width: '15%',
-    },
-    cellCreated: {
-      width: '15%',
-      minWidth: 175,
-    },
-    cellUpdated: {
-      width: '15%',
-      minWidth: 175,
-    },
-    cellUpdatedBy: {
-      width: '10%',
-      minWidth: 120,
-    },
-  });
-
-type CombinedProps = Props &
-  Omit<OrderByProps, 'data'> &
-  WithStyles<ClassNames>;
+export type CombinedProps = Props &
+  PaginationProps<SupportTicket> &
+  Omit<OrderByProps, 'data'>;
 
 export class TicketList extends React.Component<CombinedProps, {}> {
   mounted: boolean = false;
@@ -124,28 +83,26 @@ export class TicketList extends React.Component<CombinedProps, {}> {
       count,
       page,
       pageSize,
-      classes,
     } = this.props;
 
     const isActive = (label: string) => label === orderBy;
 
     return (
-      <React.Fragment>
-        <Paper>
-          <Table aria-label="List of Tickets">
-            <TableHead>
-              <TableRow>
-                <TableSortCell
-                  label="summary"
-                  direction={order}
-                  handleClick={handleOrderChange}
-                  active={isActive('summary')}
-                  data-qa-support-subject-header
-                  noWrap
-                  className={classes.cellSubject}
-                >
-                  Subject
-                </TableSortCell>
+      <>
+        <Table aria-label="List of Tickets">
+          <TableHead>
+            <TableRow>
+              <TableSortCell
+                label="summary"
+                direction={order}
+                handleClick={handleOrderChange}
+                active={isActive('summary')}
+                data-qa-support-subject-header
+                noWrap
+              >
+                Subject
+              </TableSortCell>
+              <Hidden smDown>
                 <TableSortCell
                   label="id"
                   direction={order}
@@ -153,16 +110,12 @@ export class TicketList extends React.Component<CombinedProps, {}> {
                   active={isActive('id')}
                   data-qa-support-id-header
                   noWrap
-                  className={classes.cellId}
                 >
                   Ticket ID
                 </TableSortCell>
-                <TableCell
-                  data-qa-support-regarding-header
-                  className={classes.cellRegarding}
-                >
-                  Regarding
-                </TableCell>
+              </Hidden>
+              <TableCell data-qa-support-regarding-header>Regarding</TableCell>
+              <Hidden xsDown>
                 <TableSortCell
                   label="opened"
                   direction={order}
@@ -170,7 +123,6 @@ export class TicketList extends React.Component<CombinedProps, {}> {
                   active={isActive('opened')}
                   data-qa-support-date-header
                   noWrap
-                  className={classes.cellCreated}
                 >
                   Date Created
                 </TableSortCell>
@@ -181,10 +133,11 @@ export class TicketList extends React.Component<CombinedProps, {}> {
                   active={isActive('updated')}
                   data-qa-support-updated-header
                   noWrap
-                  className={classes.cellUpdated}
                 >
                   Last Updated
                 </TableSortCell>
+              </Hidden>
+              <Hidden smDown>
                 <TableSortCell
                   label="updated_by"
                   direction={order}
@@ -192,30 +145,27 @@ export class TicketList extends React.Component<CombinedProps, {}> {
                   active={isActive('updated_by')}
                   data-qa-support-updated-by-header
                   noWrap
-                  className={classes.cellUpdatedBy}
                 >
                   Updated By
                 </TableSortCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>{this.renderContent()}</TableBody>
-          </Table>
-          <PaginationFooter
-            count={count}
-            page={page}
-            pageSize={pageSize}
-            handlePageChange={this.props.handlePageChange}
-            handleSizeChange={this.props.handlePageSizeChange}
-            eventCategory="ticket list"
-            padded
-          />
-        </Paper>
-      </React.Fragment>
+              </Hidden>
+            </TableRow>
+          </TableHead>
+          <TableBody>{this.renderContent()}</TableBody>
+        </Table>
+        <PaginationFooter
+          count={count}
+          page={page}
+          pageSize={pageSize}
+          handlePageChange={this.props.handlePageChange}
+          handleSizeChange={this.props.handlePageSizeChange}
+          eventCategory="ticket list"
+          padded
+        />
+      </>
     );
   }
 }
-
-const styled = withStyles(styles);
 
 const updatedRequest = (ownProps: Props, params: any, filters: any) => {
   return getTicketsPage(params, filters, ownProps.filterStatus);
@@ -223,4 +173,4 @@ const updatedRequest = (ownProps: Props, params: any, filters: any) => {
 
 const paginated = Pagey(updatedRequest);
 
-export default compose(paginated, styled)(TicketList);
+export default compose<CombinedProps, Props>(paginated)(TicketList);
