+++ packages/manager/src/features/Support/SupportTickets/TicketRow.tsx
@@ -1,6 +1,7 @@
 import { SupportTicket } from '@linode/api-v4/lib/support';
 import * as React from 'react';
 import { Link } from 'react-router-dom';
+import Hidden from 'src/components/core/Hidden';
 import {
   createStyles,
   withStyles,
@@ -8,8 +9,8 @@ import {
 } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
 import DateTimeDisplay from 'src/components/DateTimeDisplay';
-import TableCell from 'src/components/TableCell';
-import TableRow from 'src/components/TableRow';
+import TableCell from 'src/components/TableCell/TableCell_CMR';
+import TableRow from 'src/components/TableRow/TableRow_CMR';
 import { getLinkTargets } from 'src/utilities/getEventsActionLink';
 
 interface Props {
@@ -58,32 +59,31 @@ const TicketRow: React.FC<CombinedProps> = (props) => {
       data-testid="ticket-row"
       ariaLabel={`Ticket subject ${ticket.summary}`}
     >
-      <TableCell parentColumn="Subject" data-qa-support-subject>
+      <TableCell data-qa-support-subject>
         <Link to={`/support/tickets/${ticket.id}`}>
           <Typography variant="h3" className={classes.summary}>
             {ticket.summary}
           </Typography>
         </Link>
       </TableCell>
-      <TableCell parentColumn="Ticket ID" data-qa-support-id>
-        {ticket.id}
-      </TableCell>
-      <TableCell
-        parentColumn="Regarding"
-        data-qa-support-entity
-        className={classes.regarding}
-      >
+      <Hidden smDown>
+        <TableCell data-qa-support-id>{ticket.id}</TableCell>
+      </Hidden>
+
+      <TableCell data-qa-support-entity className={classes.regarding}>
         {renderEntityLink(ticket)}
       </TableCell>
-      <TableCell parentColumn="Date Created" data-qa-support-date>
-        <DateTimeDisplay value={ticket.opened} />
-      </TableCell>
-      <TableCell parentColumn="Last Updated" data-qa-support-updated>
-        <DateTimeDisplay value={ticket.updated} />
-      </TableCell>
-      <TableCell parentColumn="Updated By" data-qa-support-updated-by>
-        {ticket.updated_by}
-      </TableCell>
+      <Hidden xsDown>
+        <TableCell data-qa-support-date>
+          <DateTimeDisplay value={ticket.opened} />
+        </TableCell>
+        <TableCell data-qa-support-updated>
+          <DateTimeDisplay value={ticket.updated} />
+        </TableCell>
+      </Hidden>
+      <Hidden smDown>
+        <TableCell data-qa-support-updated-by>{ticket.updated_by}</TableCell>
+      </Hidden>
     </TableRow>
   );
 };
